requirements.txt
pdfminer.six
pyfiglet
